refactor: add feature flags
