+++ iroh/src/database.rs
@@ -1,4 +1,5 @@
-//! Various database implementations for storing blob data, to be used with
-//! [crate::node::Node].
+//! Various database implementations for storing blob data
+#[cfg(feature = "flat-db")]
 pub mod flat;
+#[cfg(feature = "mem-db")]
 pub mod mem;
